created utils file to hold greeting and fetching first name functions so that it can be reusable in other files
